+++ pylabrobot/resources/__init__.py
@@ -17,12 +17,10 @@ from .celltreat import *
 from .cellvis import *
 from .container import Container
 from .coordinate import Coordinate
-from .corning_axygen import *
-from .corning_costar import *
+from .corning import *
 from .deck import Deck
 from .eppendorf import *
 from .errors import ResourceNotFoundError
-from .falcon import *
 from .hamilton import *
 from .itemized_resource import ItemizedResource
 from .liquid import Liquid
