+++ pylabrobot/resources/corning/costar/__init__.py
@@ -1,0 +1,1 @@
+from .plates import *
