+++ pylabrobot/resources/corning_axygen/__init__.py
@@ -1,1 +1,3 @@
-from .plates import *
+raise NotImplementedError(
+  "pylabrobot.resources.corning_axygen is deprecated. Please use pylabrobot.resources.corning.axygen instead."
+)
