+++ pylabrobot/resources/corning_costar/__init__.py
@@ -1,1 +1,3 @@
-from .plates import *
+raise NotImplementedError(
+  "pylabrobot.resources.corning_costar is deprecated. Please use pylabrobot.resources.corning.costar instead."
+)
