+++ pylabrobot/resources/falcon/__init__.py
@@ -1,1 +1,3 @@
-from .tubes import *
+raise NotImplementedError(
+  "pylabrobot.resources.falcon is deprecated. Please use pylabrobot.resources.corning.falcon instead."
+)
